Change race filtering in match analysis to be more inclusive *
You can now search for any number of factions. Previously searching for more than two factions in 1v1 games would give no results because all of the factions were required to be included. Now the filtering is more permissive and includes results as long as the participating factions were included in the filter.